myproject/requirements.txt
dotenv
requests
django-cors-headers